feat: add image cropper, fix photo auth, remove confidence
- Add crop + rotate editor for receipt uploads (Canvas-based)
- Fix photo endpoint to work with both receipts and proposals
- Fix photo loading to use blob URLs with auth headers
- Add 'Open in new tab' link for full resolution viewing
- Add zoom/pan for receipt images in proposal view
- Remove confidence from UI, domain, and proto
- Support fractional quantities for weighted items (float64)
- Disable LLM thinking for receipt parsing
- Store error message for failed proposals
- Add cache/ to gitignore

+++ internal/api/proposals.go
@@ -196,8 +196,8 @@ func (r *Router) handleDeleteProposal(w http.ResponseWriter, req *http.Request) 
 }
 
 type updateProposalItemRequest struct {
-	ParsedName     string `json:"parsedName"`
-	Quantity       uint32 `json:"quantity"`
+	ParsedName     string  `json:"parsedName"`
+	Quantity       float64 `json:"quantity"`
 	UnitPriceCents int64  `json:"unitPriceCents"`
 }
